+++ src/app/constants.js
@@ -1,3 +1,8 @@
+/**
+ * This file defines the current constant values.
+ * To deprecate a value, add it to ./constants-old.js
+ * with a corresponding log message.
+ */
 export const Component = {
   DESCRIPTION: 'description',
   STATUS: 'status',
@@ -22,6 +27,7 @@ export const DataType = {
   NEIGHBORHOODS: 'neighborhoods',
   RASTER: 'raster',
 };
+
 
 export const FileType = {
   CELLS_JSON: 'cells.json',
@@ -85,10 +91,10 @@ export const CoordinationType = {
   GENE_EXPRESSION_TRANSFORM: 'geneExpressionTransform',
   GENE_EXPRESSION_COLORMAP_RANGE: 'geneExpressionColormapRange',
   CELL_COLOR_ENCODING: 'cellColorEncoding',
-  SPATIAL_RASTER_LAYERS: 'spatialRasterLayers',
-  SPATIAL_CELLS_LAYER: 'spatialCellsLayer',
-  SPATIAL_MOLECULES_LAYER: 'spatialMoleculesLayer',
-  SPATIAL_NEIGHBORHOODS_LAYER: 'spatialNeighborhoodsLayer',
+  SPATIAL_RASTER_LAYER: 'spatialRasterLayer',
+  SPATIAL_SEGMENTATION_LAYER: 'spatialSegmentationLayer',
+  SPATIAL_POINT_LAYER: 'spatialPointLayer',
+  SPATIAL_NEIGHBORHOOD_LAYER: 'spatialNeighborhoodLayer',
   GENOMIC_ZOOM_X: 'genomicZoomX',
   GENOMIC_ZOOM_Y: 'genomicZoomY',
   GENOMIC_TARGET_X: 'genomicTargetX',