+++ src/app/view-config-versions.js
@@ -11,7 +11,8 @@ import configSchema1_0_6 from '../schemas/config-1.0.6.schema.json';
 import configSchema1_0_7 from '../schemas/config-1.0.7.schema.json';
 import configSchema1_0_8 from '../schemas/config-1.0.8.schema.json';
 import configSchema1_0_9 from '../schemas/config-1.0.9.schema.json';
-// TODO
+import configSchema1_0_10 from '../schemas/config-1.0.10.schema.json';
+import configSchema1_0_11 from '../schemas/config-1.0.11.schema.json';
 import configSchema1_0_12 from '../schemas/config-1.0.12.schema.json';
 import cellSetsSchema from '../schemas/cell-sets.schema.json';
 import rasterSchema from '../schemas/raster.schema.json';
@@ -26,7 +27,8 @@ import {
   upgradeFrom1_0_6,
   upgradeFrom1_0_7,
   upgradeFrom1_0_8,
-  // TODO
+  upgradeFrom1_0_9,
+  upgradeFrom1_0_10,
   upgradeFrom1_0_11,
 } from './view-config-upgraders';
 
@@ -47,7 +49,8 @@ export const SCHEMA_HANDLERS = {
   '1.0.6': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_6), upgradeFrom1_0_6],
   '1.0.7': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_7), upgradeFrom1_0_7],
   '1.0.8': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_8), upgradeFrom1_0_8],
-  '1.0.9': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_9), upgradeFrom1_0_11], // TODO
-  // TODO
+  '1.0.9': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_9), upgradeFrom1_0_9],
+  '1.0.10': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_10), upgradeFrom1_0_10],
+  '1.0.11': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_11), upgradeFrom1_0_11],
   '1.0.12': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_12), null],
 };