+++ src/app/vitessce-grid-layout/layout-utils.test.js
@@ -79,19 +79,23 @@ describe('layout-utils.js', () => {
 
   describe('resolveLayout', () => {
     const componentsSpec = [
-      { component: 'NoProps', x: 0, y: 0 },
       {
-        component: 'HasProps', props: { foo: 'bar' }, x: 1, y: 1, w: 1, h: 1,
+        uid: 'i0', component: 'NoProps', x: 0, y: 0,
+      },
+      {
+        uid: 'i1', component: 'HasProps', props: { foo: 'bar' }, x: 1, y: 1, w: 1, h: 1,
       },
     ];
     const expectedComponents = {
       i0: {
+        uid: 'i0',
         component: 'NoProps',
         coordinationScopes: {},
         coordinationScopesBy: {},
         props: {},
       },
       i1: {
+        uid: 'i1',
         component: 'HasProps',
         coordinationScopes: {},
         coordinationScopesBy: {},